Add constructed tables (views) to DMI
I was getting frustrated with how much work it was to get several
related records at once, so I implemented a way of specifying
groups of related records in the schema and a method of getting them
in the DMIs.

This initial implementation has the constructed tables using the generic
DMI methods, which has the advantage that support for this feature will
not have to be written for each individual adapter but the disadvantage
that it will usually take longer and make more calls (network, db, etc)
than necessary. There are two obvious directions for further
improvements:

  1) implement this in each adapter in an efficient way (e.g. using
     JOINs in the db adapter rather than a series of SELECTs)
  2) implement a generic way for DMIs to attempt to delegate stuff like
     this upstream, e.g. instead of the API DMI doing a bunch of GETs
     to resolve everything in the constructed table, have it try to do
     a GET on the constructed table name and hope that the db DMI behind
     the server will do the right thing, falling back to the first
     strategy on error.

One last consideration about this feature is that there's really no
need for the constructed table definitions to be in the schema--they
could just as easily be created at runtime, and a sufficiently smart
server could even dynamically generate endpoints for a client that
submitted a new constructed table definition. The reason I didn't
implement that was that I didn't have an immediate use for it and I
also thought it best for readability to have constructed table
definitions all collected in the schema rather than scattered
throughout client code. Still, the option is there.

+++ water-shape/persistence/sqlite3-adapter.js
@@ -3,6 +3,7 @@ const _ = require('lodash');
 const boom = require('boom');
 const sqlite3 = require('sqlite3');
 const uuid = require('uuid');
+const constructedTableFactory = require('../generic/constructed');
 
 /*
  * API for data access:
@@ -170,7 +171,11 @@ module.exports = function(filename, schema, logger, callback) {
     }
 
     function buildSqliteSchema(schema) {
-      var finishedTables = [];
+      var finishedTables = _.filter(_.map(schema, function(v, k) {
+        if (v.constructed) {
+          return k;
+        }
+      }));
       var tableDependencies = _.reduce(schema, function(acc, val, key) {
         var dependencies = [];
         var foreignKeyConstraints = _.get(val, 'constraints.FOREIGN_KEYS');
@@ -234,7 +239,7 @@ module.exports = function(filename, schema, logger, callback) {
       return db.exec(buildSqliteSchema(schema), callback);
     }
 
-    var response = {
+    var dmi = {
       schema: _.cloneDeep(schema),
       close: function(cb) {db.close(cb);},
       buildSqliteSchema: buildSqliteSchema,
@@ -246,18 +251,22 @@ module.exports = function(filename, schema, logger, callback) {
       getAllRowsFromTable: getAllRowsFromTable
     };
     _.each(schema, function(tableDescription, tableName) {
-      var tableMethods = {};
-      tableMethods.save = _.partial(upsertIntoDb, tableName);
-      tableMethods.update = tableMethods.save;
-      tableMethods.delete = _.partial(remove, tableName);
-      tableMethods.deleteById = _.partial(removeById, tableName);
-      tableMethods.list = _.partial(getAllRowsFromTable, tableName);
-      tableMethods.getById = _.partial(getFromDbById, tableName);
-      tableMethods.search = _.partial(searchInTable, tableName);
-      response[tableName] = tableMethods;
+      if (!tableDescription.constructed) {
+        var tableMethods = {};
+        tableMethods.save = _.partial(upsertIntoDb, tableName);
+        tableMethods.update = tableMethods.save;
+        tableMethods.delete = _.partial(remove, tableName);
+        tableMethods.deleteById = _.partial(removeById, tableName);
+        tableMethods.list = _.partial(getAllRowsFromTable, tableName);
+        tableMethods.getById = _.partial(getFromDbById, tableName);
+        tableMethods.search = _.partial(searchInTable, tableName);
+        dmi[tableName] = tableMethods;
+      } else {
+        dmi[tableName] = constructedTableFactory.createConstructedTable(dmi, tableDescription, tableName);
+      }
     });
     db.run('PRAGMA foreign_keys = ON;', function() {
-      callback(response);
+      callback(dmi);
     });
   }
 }
